requirements.txt
chromadb==0.5.15
empy==3.3.4
httpx==0.27.2
huggingface-hub==0.26.1
gTTS==2.5.2
keyboard==0.13.5
langchain==0.2.16
langchain-chroma==0.1.4
langchain-community==0.2.17
langchain-core==0.2.41
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.4
langsmith==0.1.136
numpy==1.24.4
openai==1.70.0
opentelemetry-api==1.27.0
playsound==1.3.0
protobuf==3.20.3
PyAudio==0.2.14 
pydub==0.25.1
pygame==2.5.2
pypdf==5.1.0
pysqlite3-binary==0.5.4
scipy==1.10.1
sentence-transformers==3.2.1
sounddevice==0.4.7
SpeechRecognition==3.10.4
SQLAlchemy==2.0.36
vosk==0.3.45
tokenizers==0.20.1
transformers==4.46.0
